Managing the Default Heap Using RPG Operations
The first request for dynamic storage within an activation group results
in the creation of a default heap from which the storage allocation
takes place. Additional requests for dynamic storage are met by further
allocations from the default heap. If there is insufficient storage in
the heap to satisfy the current request for dynamic storage, the heap is
extended and the additional storage is allocated.

Allocated dynamic storage remains allocated until it is explicitly freed
or until the heap is discarded. The default heap is discarded only when
the owning activation group ends.

Programs in the same activation group all use the same default heap. If
one program accesses storage beyond what has be allocated, it can cause
problems for another program. For example, assume that two programs, PGM
A and PGM B are running in the same activation group. 10 bytes are
allocated for PGM A, but 11 bytes are changed by PGM A. If the extra
byte was in fact allocated for PGM B, problems may arise for PGM B.

You can use the following RPG operations on the default heap:
The ALLOC operation code and the %ALLOC built-in function allocate
storage within the default heap.
The DEALLOC operation code frees one previous allocation of heap storage
from any heap.
The REALLOC operation code and the %REALLOC built-in function change the
size of previously allocated storage from any heap.
Note
Although ALLOC and %ALLOC work only with the default heap, DEALLOC,
REALLOC, and %REALLOC work with both the default heap and user-created
heaps.